+++ src/zoid/saved-payment-methods/container.jsx
@@ -91,7 +91,6 @@ export function containerTemplate({
             top: 0;
             left: 0;
             width: 100%;
-            height: 100%;
             border: none;
           }